# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Riksdagskollen API
  description: Tracks Swedish politicians' stated promises and party goals against their actual voting records in Riksdagen.
  version: "1.0.0"
  contact:
    name: Riksdagskollen
servers:
  - url: /api
    description: API base path
...
paths:
...
  /parties/{party}/goals:
    get:
      operationId: listPartyGoals
      tags: [parties]
      summary: List all goals for a party with alignment percentages
      parameters:
        - $ref: "#/components/parameters/PartyParam"
        - name: topic
          in: query
          schema:
            $ref: "#/components/schemas/Topic"
      responses:
        "200":
          description: Party goals with vote alignment data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GoalWithAlignment"
        "404":
          $ref: "#/components/responses/NotFound"
  /parties/{party}/goals/{goalId}/votes:
    get:
      operationId: listGoalVotes
      tags: [parties]
      summary: Detailed vote breakdown for a specific party goal
      parameters:
        - $ref: "#/components/parameters/PartyParam"
        - name: goalId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Votes relevant to this goal with context notes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalVoteBreakdown"
        "404":
          $ref: "#/components/responses/NotFound"
  /politicians:
    get:
      operationId: listPoliticians
      tags: [politicians]
      summary: List politicians with summary statistics
      parameters:
        - name: party
          in: query
          schema:
            $ref: "#/components/schemas/PartyCode"
        - name: active
          in: query
          schema:
            type: boolean
            default: true
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Paginated list of politicians
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoliticianListResponse"
  /politicians/{id}:
    get:
      operationId: getPolitician
      tags: [politicians]
      summary: Full politician profile
      parameters:
        - $ref: "#/components/parameters/PoliticianIdParam"
      responses:
        "200":
          description: Politician profile with statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoliticianProfile"
        "404":
          $ref: "#/components/responses/NotFound"
  /politicians/{id}/promises:
    get:
      operationId: listPoliticianPromises
      tags: [politicians]
      summary: Promises extracted from speeches with vote match results
      parameters:
        - $ref: "#/components/parameters/PoliticianIdParam"
        - name: topic
          in: query
          schema:
            $ref: "#/components/schemas/Topic"
      responses:
        "200":
          description: Promises with vote alignment
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PromiseWithMatches"
        "404":
          $ref: "#/components/responses/NotFound"
  /politicians/{id}/votes:
    get:
      operationId: listPoliticianVotes
      tags: [politicians]
      summary: Voting record with proposal origin context
      parameters:
        - $ref: "#/components/parameters/PoliticianIdParam"
        - name: session
          in: query
          schema:
            type: string
            example: "2024/25"
        - name: topic
          in: query
          schema:
            $ref: "#/components/schemas/Topic"
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Voting record with proposal origin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteListResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /votes/{beteckning}/{punkt}:
    get:
      operationId: getVoteDetail
      tags: [votes]
      summary: Full vote detail with proposal origin and all party positions
      parameters:
        - name: beteckning
          in: path
          required: true
          schema:
            type: string
            example: "SoU12"
        - name: punkt
          in: path
          required: true
          schema:
            type: string
            example: "1"
      responses:
        "200":
          description: Full vote detail with party breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoteDetail"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /context/topic/{topic}:
    get:
      operationId: getTopicContext
      tags: [context]
      summary: Budget and economic context for a political topic
      parameters:
        - name: topic
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Topic"
      responses:
        "200":
          description: Topic context with budget trend and funding analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopicContext"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    PartyParam:
      name: party
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/PartyCode"
    PoliticianIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        description: Riksdagen intressent_id
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        detail:
          type: string
    PartyCode:
      type: string
      enum: [S, M, SD, C, V, KD, L, MP, unknown]
      description: Swedish party abbreviation
    Topic:
      type: string
      enum:
        - sjukvard
        - forsvar
        - invandring
        - skatt
        - skola
        - klimat
        - brott
        - bostad
        - arbete
        - energi
        - other
    VoteResult:
      type: string
      enum: [Ja, Nej, Avstår, Frånvarande]
    ProposalType:
      type: string
      enum: [prop, mot, bet]
      description: "prop=government bill, mot=party motion, bet=committee report"
    Specificity:
      type: string
      enum: [concrete, directional, rhetorical]
    Politician:
      type: object
      required: [intressentId, firstName, lastName, party, isActive]
      properties:
        intressentId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        party:
          $ref: "#/components/schemas/PartyCode"
        constituency:
          type: string
        imageUrl:
          type: string
          format: uri
        isActive:
          type: boolean
    PoliticianSummary:
      allOf:
        - $ref: "#/components/schemas/Politician"
        - type: object
          properties:
            totalVotes:
              type: integer
            promiseCount:
              type: integer
            alignmentPct:
              type: number
              format: float
              description: Overall promise-to-vote alignment percentage
    PoliticianProfile:
      allOf:
        - $ref: "#/components/schemas/PoliticianSummary"
        - type: object
          properties:
            recentSpeeches:
              type: array
              items:
                $ref: "#/components/schemas/SpeechSummary"
    PoliticianListResponse:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PoliticianSummary"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    SpeechSummary:
      type: object
      required: [id, date, topicHeading]
      properties:
        id:
          type: integer
        dokId:
          type: string
        date:
          type: string
          format: date
        topicHeading:
          type: string
    Vote:
      type: object
      required: [id, voteringId, politicianId, party, voteResult, beteckning, forslagspunkt, session]
      properties:
        id:
          type: integer
        voteringId:
          type: string
        politicianId:
          type: string
        party:
          $ref: "#/components/schemas/PartyCode"
        voteResult:
          $ref: "#/components/schemas/VoteResult"
        beteckning:
          type: string
          description: Committee report ID e.g. SoU12
        forslagspunkt:
          type: string
        session:
          type: string
          example: "2024/25"
        dokId:
          type: string
        proposedByParty:
          $ref: "#/components/schemas/PartyCode"
        proposalType:
          $ref: "#/components/schemas/ProposalType"
        documentTitle:
          type: string
        createdAt:
          type: string
          format: date-time
    VoteListResponse:
      type: object
      required: [data, total, page, pageSize]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Vote"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
    VoteDetail:
      type: object
      required: [beteckning, forslagspunkt, documentTitle, partyBreakdown]
      properties:
        beteckning:
          type: string
        forslagspunkt:
          type: string
        dokId:
          type: string
        documentTitle:
          type: string
        proposedByParty:
          $ref: "#/components/schemas/PartyCode"
        proposalType:
          $ref: "#/components/schemas/ProposalType"
        session:
          type: string
        partyBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/PartyVotePosition"
        contextNote:
          type: string
          description: AI-generated explanation of political context
    PartyVotePosition:
      type: object
      required: [party, jaCount, nejCount, avstarCount, franvarandeCount]
      properties:
        party:
          $ref: "#/components/schemas/PartyCode"
        jaCount:
          type: integer
        nejCount:
          type: integer
        avstarCount:
          type: integer
        franvarandeCount:
          type: integer
        dominantVote:
          $ref: "#/components/schemas/VoteResult"
    Goal:
      type: object
      required: [id, party, goalText, topic, specificity, sourceDocument]
      properties:
        id:
          type: integer
        party:
          $ref: "#/components/schemas/PartyCode"
        goalText:
          type: string
        topic:
          $ref: "#/components/schemas/Topic"
        specificity:
          $ref: "#/components/schemas/Specificity"
        sourceDocument:
          type: string
          example: "Valmanifest 2022"
        keywords:
          type: array
          items:
            type: string
        relevantCommittees:
          type: array
          items:
            type: string
    GoalWithAlignment:
      allOf:
        - $ref: "#/components/schemas/Goal"
        - type: object
          properties:
            relevantVotes:
              type: integer
            alignmentPct:
              type: number
              format: float
              description: Percentage of relevant votes aligned with goal direction
    GoalVoteBreakdown:
      type: object
      required: [goal, matches]
      properties:
        goal:
          $ref: "#/components/schemas/GoalWithAlignment"
        matches:
          type: array
          items:
            $ref: "#/components/schemas/GoalVoteMatch"
    GoalVoteMatch:
      type: object
      required: [beteckning, forslagspunkt, relevanceScore, alignedDirection]
      properties:
        beteckning:
          type: string
        forslagspunkt:
          type: string
        relevanceScore:
          type: number
          format: float
          minimum: 0
          maximum: 1
        alignedDirection:
          $ref: "#/components/schemas/VoteResult"
        explanation:
          type: string
        proposedByParty:
          $ref: "#/components/schemas/PartyCode"
        proposalType:
          $ref: "#/components/schemas/ProposalType"
        contextNote:
          type: string
        documentTitle:
          type: string
        partyAlignmentPct:
          type: number
          format: float
          description: What % of party members voted in the aligned direction
...
    Promise:
      type: object
      required: [id, politicianId, promiseText, topic, specificity]
      properties:
        id:
          type: integer
        politicianId:
          type: string
        speechId:
          type: integer
        speechDate:
          type: string
          format: date
        speechTopic:
          type: string
        promiseText:
          type: string
        topic:
          $ref: "#/components/schemas/Topic"
        specificity:
          $ref: "#/components/schemas/Specificity"
        keywords:
          type: array
          items:
            type: string
    PromiseWithMatches:
      allOf:
        - $ref: "#/components/schemas/Promise"
        - type: object
          properties:
            voteMatches:
              type: array
              items:
                $ref: "#/components/schemas/PromiseVoteMatch"
    PromiseVoteMatch:
      type: object
      required: [voteId, relevanceScore, alignment]
      properties:
        voteId:
          type: integer
        beteckning:
          type: string
        forslagspunkt:
          type: string
        documentTitle:
          type: string
        voteResult:
          $ref: "#/components/schemas/VoteResult"
        relevanceScore:
          type: number
          format: float
        alignment:
          type: string
          enum: [supports, contradicts, unclear]
        explanation:
          type: string
        proposedByParty:
          $ref: "#/components/schemas/PartyCode"
        proposalType:
          $ref: "#/components/schemas/ProposalType"
...
    BudgetTrendEntry:
      type: object
      required: [year, amountKsek, amountFormatted]
      properties:
        year:
          type: integer
        amountKsek:
          type: integer
          format: int64
        amountFormatted:
          type: string
          description: "Human-readable amount, e.g. '87,3 mdkr'"
    RelatedBudgetArea:
      type: object
      required: [code, name, trend, latestDeltaPct, shareOfBudgetPct]
      properties:
        code:
          type: string
          example: "UO9"
        name:
          type: string
        trend:
          type: array
          items:
            $ref: "#/components/schemas/BudgetTrendEntry"
        latestDeltaPct:
          type: number
          format: float
          description: Year-over-year change in percent
        shareOfBudgetPct:
          type: number
          format: float
          description: Area's share of total national budget in the latest year
    FundingRow:
      type: object
      required: [code, name, deltaKsek, deltaPct]
      properties:
        code:
          type: string
        name:
          type: string
        deltaKsek:
          type: integer
          format: int64
        deltaPct:
          type: number
          format: float
    FundingContext:
      type: object
      required: [year, topIncreases, topDecreases]
      properties:
        year:
          type: integer
        topIncreases:
          type: array
          items:
            $ref: "#/components/schemas/FundingRow"
        topDecreases:
          type: array
          items:
            $ref: "#/components/schemas/FundingRow"
        totalBudgetDeltaKsek:
          type: integer
          format: int64
        totalBudgetDeltaPct:
          type: number
          format: float
    TopicContext:
      type: object
      required: [topic, description, committees, relatedAreas]
      properties:
        topic:
          $ref: "#/components/schemas/Topic"
        description:
          type: string
        committees:
          type: array
          items:
            type: string
        relatedAreas:
          type: array
          items:
            $ref: "#/components/schemas/RelatedBudgetArea"
        fundingContext:
          $ref: "#/components/schemas/FundingContext"
...